reservation

+++ src/main/java/com/example/demo/MedicalRecord/MedicalRecordRepository.java
@@ -2,4 +2,4 @@ package com.example.demo.MedicalRecord;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicalRecordRepository extends JpaRepository<MedicalRecord,Long> { }
+public interface MedicalRecordRepository extends JpaRepository<MedicalRecord,Integer> { }
