+++ src/main/java/com/example/demo/reception/ReceptionRepository.java
@@ -2,4 +2,4 @@ package com.example.demo.reception;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceptionRepository extends JpaRepository<Reception,Long> { }
+public interface ReceptionRepository extends JpaRepository<Reception,Integer> { }
